Move shops UI to Houses module
