+++ src/Application/Game.cpp
@@ -69,8 +69,8 @@
 #include "GUI/UI/UIQuickReference.h"
 #include "GUI/UI/UIRest.h"
 #include "GUI/UI/UISaveLoad.h"
-#include "GUI/UI/UIShops.h"
 #include "GUI/UI/UIStatusBar.h"
+#include "GUI/UI/Houses/Shops.h"
 
 #include "Io/Mouse.h"
 #include "Io/KeyboardInputHandler.h"
