+++ src/GUI/UI/Houses/MagicGuild.cpp
@@ -17,7 +17,7 @@
 #include "GUI/GUIButton.h"
 #include "GUI/GUIFont.h"
 #include "GUI/UI/UIStatusBar.h"
-#include "GUI/UI/UIShops.h"
+#include "GUI/UI/Houses/Shops.h"
 
 #include "Io/Mouse.h"
